requirements.txt
absl-py==2.1.0
accelerate==0.29.3
aiohttp==3.9.5
aiosignal==1.3.1
anytree==2.12.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
autoawq==0.2.5
autoawq_kernels==0.0.6
beautifulsoup4==4.12.3
bert-score==0.3.13
bitsandbytes==0.43.1
boto3==1.34.89
botocore==1.34.89
bpemb==0.3.5
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
conllu==4.5.3
contourpy==1.1.1
cycler==0.12.1
datasets==2.19.0
Deprecated==1.2.14
dill==0.3.8
docopt==0.6.2
docstring_parser==0.16
editdistance==0.8.1
eval_type_backport==0.2.0
filelock==3.13.4
flair==0.13.1
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gast==0.4.0
gdown==5.1.0
gensim==4.3.2
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
huggingface-hub==0.24.6
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
Janome==0.5.0
jieba==0.42.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
keras==2.13.1
kiwisolver==1.4.5
langdetect==1.0.9
language-tool-python==2.8
lemminflect==0.2.3
libclang==18.1.1
lru-dict==1.3.0
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
more-itertools==10.2.0
mpld3==0.5.10
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.1
nltk==3.8.1
num2words==0.5.13
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
OpenHowNet==2.0
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
peft==0.10.0
pillow==10.3.0
pinyin==0.4.0
pptree==3.1
protobuf==4.25.4
psutil==5.9.8
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-ternary==1.0.8
pytorch_revgrad==0.2.0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
segtok==1.5.11
semver==3.0.2
sentencepiece==0.2.0
shtab==1.7.1
six==1.16.0
smart-open==7.0.4
soupsieve==2.5
sqlitedict==2.1.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
terminaltables==3.1.10
textattack==0.3.10
tf-keras==2.15.0
threadpoolctl==3.4.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.2
transformer-smaller-training-vocab==0.4.0
transformers==4.44.2
triton==2.2.0
trl==0.8.6
typing_extensions==4.5.0
tyro==0.8.3
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.2.13
Werkzeug==3.0.3
Wikipedia-API==0.6.0
word2number==1.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
zstandard==0.22.0
